requirements-to-freeze.txt
codecov==2.0.15
django
djangorestframework
django-filter
django-heroku
green==2.12.1
gunicorn
invoke==1.2.0
markdown
pycodestyle==2.4.0
pydocstyle==2.1.1
PyGithub
pytest==3.8.1
pytest-cov==2.6.0
radon==2.2.0
sphinx==1.8.1
sphinx-autobuild==0.7.1
sphinx_rtd_theme==0.4.1
whitenoise
